YUK-14841 Editing System Device Points Should Be Consistent  - Added support to view/edit "System" point types in database editor and from points page on web (intended only for Cap Control to date, but you can navigate there if you know the url and have cap control database edit access)

+++ yukon-client/yukon-web/src/com/cannontech/web/tools/points/service/impl/PointEditorServiceImpl.java
@@ -234,15 +234,17 @@ public class PointEditorServiceImpl implements PointEditorService {
      */
     private void saveStaleData(int pointId, StaleData staleData) {
         
-        PointPropertyValue timeProperty = new PointPropertyValue(pointId, StaleData.TIME_PROPERTY, staleData.getTime());
-        PointPropertyValue updateProperty = new PointPropertyValue(pointId, StaleData.UPDATE_PROPERTY, staleData.getUpdateStyle());
-        
-        pointPropertyValueDao.remove(timeProperty);
-        pointPropertyValueDao.remove(updateProperty);
-
-        if (staleData.isEnabled()) {
-            pointPropertyValueDao.add(timeProperty);
-            pointPropertyValueDao.add(updateProperty);
+        if (staleData != null) {
+            PointPropertyValue timeProperty = new PointPropertyValue(pointId, StaleData.TIME_PROPERTY, staleData.getTime());
+            PointPropertyValue updateProperty = new PointPropertyValue(pointId, StaleData.UPDATE_PROPERTY, staleData.getUpdateStyle());
+            
+            pointPropertyValueDao.remove(timeProperty);
+            pointPropertyValueDao.remove(updateProperty);
+    
+            if (staleData.isEnabled()) {
+                pointPropertyValueDao.add(timeProperty);
+                pointPropertyValueDao.add(updateProperty);
+            }
         }
     }
